Implement and enable Starlark `JavaPluginInfo`
The native provider translates the Starlark struct into the native class.

In order to use instances constructed both natively and via the Starlark constructor  in the same build, the `@StarlarkBuiltin` annotations were dropped, so that `Starlark.type()` resolves to `StructApi`. This is needed for type matching when provider instances are added to depsets.

PiperOrigin-RevId: 539978604
Change-Id: I8c1b3ae2746b515d68a5201c32059681d3b83c61