singlejar: fixes to build with msys2 on Windows
singlejar is a C++ tool that Bazel builds for the executor platform. This change resolves some compile-time errors on Windows with msys2.

In src/tools/singlejar/diag.h, use `##__VA_ARGS__` instead of `__VA_ARGS__`. This uses a language extension implemented in MSVC, GCC, and Clang that elides the leading comma when there are no arguments. MSVC did this implicitly without `##`, which is why this compiles now.

In src/main/util/numbers.h, include `<cstdint>` for `int32_t`.

Fixes #18632

Closes #18660.

PiperOrigin-RevId: 540463565
Change-Id: Ide389949894bdaeffc8771dd1682ad0df1e79330